requirements.txt
# Core MCP Framework
fastmcp>=0.2.0

# Database
aiosqlite>=0.19.0

# Fuzzy Matching
rapidfuzz>=3.6.0

# Phonetic Matching
metaphone>=0.6

# HTTP Client for data download
httpx>=0.26.0

# Data Processing
pandas>=2.1.0

# Environment Management
python-dotenv>=1.0.0

# Logging
structlog>=24.1.0